pyproject.toml
[project]
name = "fx-transaction-reporting"
version = "1.0.0"
description = ""
dependencies = [
    "flask",
    "SQLAlchemy",
    "requests",
    "PyJWT",
    "redis"
]
